Commiting before making some potentially breaking changes in order to fix portal rendering when dealing with backface culling.

+++ GameProject/Scene.cs
@@ -42,7 +42,7 @@ namespace Game
         
         public void Step()
         {
-            Step(1/Controller.StepsPerSecond);
+            Step(1/(float)Controller.StepsPerSecond);
         }
 
         public void Step(float stepSize)
